split selectors
