+++ src/selectors/fetchMeetingHistory.js
@@ -1,0 +1,18 @@
+import {
+    getMeetingsHistoryPending,
+    getMeetingsHistorySuccess,
+    getMeetingsHistoryError
+} from 'actions';
+
+export const getMeetingsHistory = () => {
+    return async dispatch => {
+        dispatch(getMeetingsHistoryPending());
+        try {
+            const response = await fetch(process.env.REACT_APP_SERVER_URL + "/meetings/archive");
+            const data = await response.json();
+            dispatch(getMeetingsHistorySuccess(data));
+        } catch(error){
+            dispatch(getMeetingsHistoryError(error));
+        };
+    };
+};
