+++ src/selectors/fetchSchedule.js
@@ -1,0 +1,18 @@
+import { 
+    getSchedulePending,
+    getScheduleSuccess,
+    getScheduleError
+} from 'actions';
+
+export const getSchedule = () => {
+    return async dispatch => {
+        dispatch(getSchedulePending());
+        try {
+            const response = await fetch(process.env.REACT_APP_SERVER_URL + "/meetings/schedule");
+            const data = await response.json();
+            dispatch(getScheduleSuccess(data));
+        } catch(error){
+            dispatch(getScheduleError(error));
+        };
+    };
+};
